% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Graphing of a country from the World Population.}
\usage{
CountryPopulation(country_name, title = NULL)
}
\arguments{
\item{a}{A real country listed}
}
\value{
A graph of the countries population each year
}
\description{
Input a country name and the function will make a graph of the countries' population each year.
}
\examples{
USA_plot <- CountryPopulation("United States of America", title = "United States of America")
}
