% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldCup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{A data set from the World Cup Wikipedia page}
\format{
a data frame with 22 observations and 4 columns
\describe{
  \item{WorldCup}{The country that hosted the world cup and corresponding year}
  \item{Total attendance}{Total people who attended}
  \item{Matches}{Number of Matches held there}
  \item{Averageattendance}{Attendance per match}
}
}
\source{
\url{www.en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
A data set from the World Cup Wikipedia page
}
\keyword{datasets}
