% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{A data set from the United Nations on Total Populations of each Country from 1950-2020}
\format{
a data frame with 16685 observations and 3 columns
\describe{
  \item{Country}{The country of interest.}
  \item{Year}{A year between 1950-2020.}
  \item{Population}{Population number}
}
}
\usage{
WorldPopulation
}
\description{
A data set from the United Nations on Total Populations of each Country from 1950-2020
}
\keyword{datasets}
